--- packages/webdoc-parser/src/build-symbol-tree/index.js
@@ -1,4 +1,0 @@
-import buildSymbolTree, {type Symbol} from "./build-symbol-tree";
-
-export default buildSymbolTree;
-export type {Symbol};
