--- packages/webdoc-parser/src/parser_.js
@@ -1,10 +1,0 @@
-// @flow
-
-import * as parser from "@babel/parser";
-import traverse, {type NodePath} from "@babel/traverse";
-
-export class Parser {
-  loadFile(file: string, fileName?: string): Symbol {
-
-  }
-}
