+++ packages/webdoc-parser/src/symbols-babel/extract-metadata.js
