+++ packages/webdoc-parser/src/types/LanguageIntegration.js
@@ -1,0 +1,7 @@
+import type {Symbol} from "./Symbol";
+
+// Language-integration definition - used for parsing code into a symbol-tree
+export type LanguageIntegration = {
+  extensions: string[],
+  parse(file: string): Symbol
+};
